@@ -13,12 +13,12 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "framer-motion": "^11.12.0",
-    "next": "13.4.12",
+    "framer-motion": "^11.13.1",
+    "next": "15.0.3",
     "next-seo": "^6.6.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-icons": "^5.3.0",
+    "react-icons": "^5.4.0",
     "react-intersection-observer": "^9.13.1"
   },
   "devDependencies": {
@@ -28,8 +28,8 @@
     "@types/node": "22.10.1",
     "@types/react": "18.3.12",
     "@types/react-dom": "18.3.1",
-    "@typescript-eslint/eslint-plugin": "^8.16.0",
-    "@typescript-eslint/parser": "^8.16.0",
+    "@typescript-eslint/eslint-plugin": "^8.17.0",
+    "@typescript-eslint/parser": "^8.17.0",
     "commitizen": "^4.3.1",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^9.16.0",
@@ -38,7 +38,7 @@
     "eslint-config-next": "15.0.3",
     "eslint-config-prettier": "^9.0.0",
     "eslint-config-react": "^1.1.7",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-cypress": "^4.1.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest-dom": "^5.5.0",
@@ -53,16 +53,16 @@
     "jest-axe": "^9.0.0",
     "jest-environment-jsdom": "^29.6.2",
     "lint-staged": "^15.2.10",
-    "prettier": "^3.4.1",
+    "prettier": "^3.4.2",
     "prettier-stylelint": "^0.4.2",
-    "sass": "^1.81.0",
+    "sass": "^1.82.0",
     "start-server-and-test": "^2.0.8",
     "stylelint": "^16.11.0",
     "stylelint-a11y": "^1.2.3",
     "stylelint-config-standard": "^36.0.1",
     "stylelint-order": "^6.0.3",
     "typescript": "5.7.2",
-    "typescript-eslint": "^8.16.0"
+    "typescript-eslint": "^8.17.0"
   },
   "config": {
     "commitizen": {
