@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 
 import { AnimatePresence, motion } from 'framer-motion';
 
@@ -13,6 +13,7 @@ const Loader = ({ isLoading, setIsLoading }: Props) => {
       setIsLoading();
     }, 1900);
   }, [setIsLoading]);
+
   return (
     <AnimatePresence>
       {isLoading && (
@@ -63,11 +64,11 @@ const Loader = ({ isLoading, setIsLoading }: Props) => {
                 strokeLinecap='round'
                 strokeLinejoin='round'
                 d='M 50, 5
-          L 11, 27
-          L 11, 72
-          L 50, 95
-          L 89, 73
-          L 89, 28 z'
+                  L 11, 27
+                  L 11, 72
+                  L 50, 95
+                  L 89, 73
+                  L 89, 28 z'
               />
             </g>
           </motion.svg>
