@@ -4,9 +4,9 @@ import Head from 'next/head';
 
 import { NextSeo } from 'next-seo';
 
-import Email from '@/components/Email';
-import Loader from '@/components/Loader';
-import SocialIcons from '@/components/SocialIcons';
+import Email from '@/components/ui/Email';
+import Loader from '@/components/ui/Loader';
+import SocialIcons from '@/components/ui/SocialIcons';
 
 import About from '@/sections/About';
 import Contact from '@/sections/Contact';
