@@ -1,8 +1,6 @@
-import React from 'react';
-
 import { motion } from 'framer-motion';
 
-import Button from '@/components/Button';
+import Button from '@/components/ui/Button';
 
 const Contact = () => {
   return (
